dts: st: remove always-on property for v3v3 on stm32mp215f-dk board
For stm32mp215f-dk board, the v3v3 is exported with a scmi regulator
and is managed by each device as ethernet phy, audio codec or display.

So remove the unnecessary regulator property always-on and the
management with power control 1, but continue to force the suspend
state in deepest low power mode, to avoid over consumption if the
application don't correctly manage this resource.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I1944d2e43d7ccc20d4f3690e454f8a66d3370aab
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/449840
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>