module: dvfs: call set_rate when frequency change with same voltage
Always set frequency with set_rate ops when the requested voltage is
equal to the actual voltage and when frequency change.

This patch allow to support several OPP in DVFS module with a same voltage
but with several frequencies.

This patch still manage the case ctx->current_opp.frequency = 0, when the
frequency has not yet been set.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I30354386edbf3feff0ff0064f68c3acfac1110e5
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/SCP-firmware/+/451573
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Valentin CARON <valentin.caron@foss.st.com>