drivers: regulator: get supply voltages as fallback
When the ops regulator_get_voltage and regulator_supported_voltages
are absent and when regulator supply is provided, the regulor should
report to voltage of their supply because the regulator is assumed to be a
simple switch  (for example STPMIC1 boost/pwr-sw1/pwr-sw2) so use ops on
this supply as fallback.

If a specific behavior for a regulator is required, the ops must
be present to implement it.

Change-Id: I619ec247b7d240eff89ff56a90b391764658f3be
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/385761
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Thomas BOURGOIN <thomas.bourgoin@foss.st.com>