product/optee-stm32mp2: add OPP backend for DVFS support
Add a module STM32 OPP as backend of CPU DVFS based on SCP clock API.
The clock API level is the minimal level to use the OP-TEE OPP driver
with SCP DVFS module as we can't implement a specific DVFS driver
in SCP firmware.

As the voltage is managed in OP-TEE OPP driver, the PSU required for
DVFS module is stubbed with a OP-TEE PSU with a NULL regulator.

NB: the OP-TEE clock module can't be used here, based on a OP-TEE OPP
    clock, because this module use directly the OP-TEE clock framework
    which add too many restrictions (it is no  allowed to call clock API
    in  clock ops).

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Signed-off-by: Valentin Caron <valentin.caron@foss.st.com>
Change-Id: I73402b8fb12d02d1e269facc98cb32900ba2159b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/SCP-firmware/+/477826
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>