product/optee-stm32mp2: update psu optee regulator module for CPU DVFS
Make adaptations to the psu optee regulator module to accept a NULL
regulator in DVFS configuration when the PSU back-end is not used for
DVFS support, for example with the  STM32 OPP module. As PSU is required
by the DVFS module a fake PSU with NULL regulator is a local PSU based
on a local context.

Moreover as regulator is always defined when trace is displayed, so
psu_regulator_name() can be removed and replaced to direct call of
OP-TEE API regulator_name() to simplify the code.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Signed-off-by: Valentin Caron <valentin.caron@foss.st.com>
Change-Id: I7f8048a9d8f5054eac8e50c06cff16b642f3ee1a
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/SCP-firmware/+/477825
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>